fix: localize errors via persisted keys, stabilize tests, finish polish
- Persist a catalog errorKey alongside the localized message so failed
  downloads re-localize after a language switch; migrate legacy English
  errorMessage entries on load and render through DownloadErrorText in the
  row, notifications and duplicate dialogs.
- Fill missing Localizable.xcstrings entries (HTTP %ld, Invalid URL,
  Version %@ (build %@), OK) and route the Settings "Launch at Login"
  alert title and button through the localization manager.
- Serialize shared state in DownloadNotifier, DownloadEngineCoordinator
  and SandboxAccess, and move global observers into ContentViewModel
  startAppServices() so the parallel test host no longer crashes;
  re-enable parallel app tests in CI.
- Centralize test-host detection in ProcessInfo.isRunningTests.
- Fix FileTypeFilter "Other" swallowing known types (fonts, disk images,
  exe/msi, torrent) and fill missing fileTypeColor cases; fix %lld Int
  mismatches; wire the new-download folder picker; drop a duplicate
  Foundation import; ship the missing multi-size app icon.

+++ MacDL/Features/Content/DialogPresenter.swift
@@ -87,7 +87,7 @@ enum DialogPresenter {
     /// Bulk delete with a "also remove files" checkbox.
     static func confirmBulkDelete(count: Int) -> (confirmed: Bool, deleteFiles: Bool) {
         let alert = NSAlert()
-        alert.messageText = String(format: LanguageManager.shared.localized("Are you sure you want to delete %lld download(s)?"), count)
+        alert.messageText = String(format: LanguageManager.shared.localized("Are you sure you want to delete %lld download(s)?"), Int64(count))
         alert.alertStyle = .warning
         alert.addButton(withTitle: LanguageManager.shared.localized("Delete"))
         alert.addButton(withTitle: LanguageManager.shared.localized("Cancel"))
